added support for postgresql, which is now the main db. the support for mysql is dropped.
added support for flyway: database schema migration tool.
the tool is run on startup of the postgresql persistence factory.

the builder does not include any configuation files into the distribution.
this is going to be changed, when we will come up with a way to keep development and production-template
files separete.